feat(payment): PI-1501 [Klarna] [CheckoutSDK] Refactor the existing `klarna` payment strategy in Checkout SDK to use the new checkout payment integration JS API
