+++ packages/core/src/payment/create-payment-strategy-registry.spec.ts
@@ -31,7 +31,6 @@ import { CyberSourceV2PaymentStrategy } from './strategies/cybersourcev2';
 import { DigitalRiverPaymentStrategy } from './strategies/digitalriver';
 import { GooglePayPaymentStrategy } from './strategies/googlepay';
 import { HummPaymentStrategy } from './strategies/humm';
-import { KlarnaPaymentStrategy } from './strategies/klarna';
 import { MasterpassPaymentStrategy } from './strategies/masterpass';
 import { OpyPaymentStrategy } from './strategies/opy';
 import { PaypalExpressPaymentStrategy, PaypalProPaymentStrategy } from './strategies/paypal';
@@ -194,12 +193,6 @@ describe('CreatePaymentStrategyRegistry', () => {
         const paymentStrategy = registry.get(PaymentStrategyType.HUMM);
 
         expect(paymentStrategy).toBeInstanceOf(HummPaymentStrategy);
-    });
-
-    it('can instantiate klarna', () => {
-        const paymentStrategy = registry.get(PaymentStrategyType.KLARNA);
-
-        expect(paymentStrategy).toBeInstanceOf(KlarnaPaymentStrategy);
     });
 
     it('can instantiate openpay', () => {
