+++ packages/core/src/payment/payment-request-options.ts
@@ -11,8 +11,6 @@ import {
 import { ChasePayInitializeOptions } from './strategies/chasepay';
 import { DigitalRiverPaymentInitializeOptions } from './strategies/digitalriver';
 import { GooglePayPaymentInitializeOptions } from './strategies/googlepay';
-import { KlarnaPaymentInitializeOptions } from './strategies/klarna';
-import { KlarnaV2PaymentInitializeOptions } from './strategies/klarnav2';
 import { MasterpassPaymentInitializeOptions } from './strategies/masterpass';
 import { MonerisPaymentInitializeOptions } from './strategies/moneris';
 import { OpyPaymentInitializeOptions } from './strategies/opy';
@@ -83,18 +81,6 @@ export interface BasePaymentInitializeOptions extends PaymentRequestOptions {
      * They can be omitted unless you need to support Digital River.
      */
     digitalriver?: DigitalRiverPaymentInitializeOptions;
-
-    /**
-     * The options that are required to initialize the Klarna payment method.
-     * They can be omitted unless you need to support Klarna.
-     */
-    klarna?: KlarnaPaymentInitializeOptions;
-
-    /**
-     * The options that are required to initialize the KlarnaV2 payment method.
-     * They can be omitted unless you need to support KlarnaV2.
-     */
-    klarnav2?: KlarnaV2PaymentInitializeOptions;
 
     /**
      * The options that are required to initialize the Masterpass payment method.
