--- packages/core/src/payment/strategies/klarna/index.ts
@@ -1,4 +1,0 @@
-export { default as KlarnaPaymentStrategy } from './klarna-payment-strategy';
-export { default as KlarnaPaymentInitializeOptions } from './klarna-payment-initialize-options';
-export { default as KlarnaCredit, KlarnaLoadResponse } from './klarna-credit';
-export { default as KlarnaScriptLoader } from './klarna-script-loader';
