--- packages/core/src/payment/strategies/klarnav2/index.ts
@@ -1,5 +1,0 @@
-export { default as KlarnaV2PaymentInitializeOptions } from './klarnav2-payment-initialize-options';
-export {
-    supportedCountries,
-    supportedCountriesRequiringStates,
-} from './klarna-supported-countries';
