+++ packages/klarna-integration/src/index.ts
@@ -1,1 +1,2 @@
+export { default as createKlarnaPaymentStrategy } from './klarna/create-klarna-payment-strategy';
 export { default as createKlarnaV2PaymentStrategy } from './klarnav2/create-klarnav2-payment-strategy';
