+++ packages/klarna-integration/src/klarna/create-klarna-payment-strategy.spec.ts
@@ -1,0 +1,20 @@
+import { PaymentIntegrationService } from '@bigcommerce/checkout-sdk/payment-integration-api';
+import { PaymentIntegrationServiceMock } from '@bigcommerce/checkout-sdk/payment-integrations-test-utils';
+
+import createKlarnaPaymentStrategy from './create-klarna-payment-strategy';
+import KlarnaPaymentStrategy from './klarna-payment-strategy';
+
+describe('createKlarnaPaymentStrategy', () => {
+    let paymentIntegrationService: PaymentIntegrationService;
+
+    beforeEach(() => {
+        paymentIntegrationService = new PaymentIntegrationServiceMock();
+    });
+
+    it('instantiates klarna payment strategy', () => {
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+        const strategy = createKlarnaPaymentStrategy(paymentIntegrationService);
+
+        expect(strategy).toBeInstanceOf(KlarnaPaymentStrategy);
+    });
+});
