+++ packages/klarna-integration/src/klarna/klarna-window.ts
@@ -1,7 +1,8 @@
+/* eslint-disable @typescript-eslint/naming-convention */
 import KlarnaCredit from './klarna-credit';
 
 export default interface KlarnaWindow extends Window {
-    Klarna: {
+    Klarna?: {
         Credit: KlarnaCredit;
     };
 }
